# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Stellar MarketPay API",
    "version": "1.0.0",
    "description": "Backend API for Stellar MarketPay - A decentralized freelance marketplace built on the Stellar blockchain.\n\n## Authentication\nMost endpoints require JWT authentication obtained via the **/api/auth** challenge-response flow using Stellar wallet signatures.\n\n## Pagination\nList endpoints support cursor-based pagination via the `after` query parameter. Responses include `next_cursor` and `has_more` fields.",
    "contact": {
      "name": "Stellar MarketPay Team",
      "email": "support@stellarmarketpay.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Development server"
    },
    {
      "url": "https://api.stellarmarketpay.com",
      "description": "Production server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /api/auth (Stellar wallet signature)"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt",
        "description": "JWT cookie set after authentication"
      },
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Developer API key for public API endpoints"
      }
    },
...
  },
  "paths": {
...
    "/api/tokens/search": {
      "get": {
        "summary": "Search tokens by name or symbol",
        "tags": [
          "Tokens"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results"
          },
          "400": {
            "description": "Missing search query"
          }
        }
      }
    },
...
    "/api/saved-searches": {
      "get": {
        "summary": "List saved searches",
        "tags": [
          "Saved Searches"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Saved searches list"
          }
        }
      },
      "post": {
        "summary": "Save a new search (max 10)",
        "tags": [
          "Saved Searches"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "query_params"
                ],
                "properties": {
                  "query_params": {
                    "type": "object"
                  },
                  "notify_in_app": {
                    "type": "boolean"
                  },
                  "notify_email": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Search saved"
          },
          "400": {
            "description": "Limit reached or invalid params"
          }
        }
      }
    },
    "/api/saved-searches/{id}": {
      "patch": {
        "summary": "Update saved search notification prefs",
        "tags": [
          "Saved Searches"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search updated"
          }
        }
      },
      "delete": {
        "summary": "Delete a saved search",
        "tags": [
          "Saved Searches"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search deleted"
          }
        }
      }
    },
...
    "/api/referrals/{publicKey}": {
      "get": {
        "summary": "Get referral stats and history",
        "tags": [
          "Referrals"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "publicKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Referral stats"
          },
          "403": {
            "description": "Can only view own referral data"
          }
        }
      }
    },
    "/api/referrals/register": {
      "post": {
        "summary": "Register a new referral",
        "tags": [
          "Referrals"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "referrerAddress",
                  "refereeAddress"
                ],
                "properties": {
                  "referrerAddress": {
                    "type": "string"
                  },
                  "refereeAddress": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Referral registered (or already exists)"
          }
        }
      }
    },
...
    "/api/onboarding": {
      "patch": {
        "summary": "Update onboarding progress",
        "tags": [
          "Onboarding"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "publicKey"
                ],
                "properties": {
                  "publicKey": {
                    "type": "string"
                  },
                  "currentStep": {
                    "type": "integer"
                  },
                  "completedSteps": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "dismissed": {
                    "type": "boolean"
                  },
                  "completed": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Onboarding progress updated"
          },
          "400": {
            "description": "Missing publicKey"
          }
        }
      }
    },
    "/api/onboarding/{publicKey}": {
      "get": {
        "summary": "Get onboarding progress for a user",
        "tags": [
          "Onboarding"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "publicKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Onboarding progress retrieved (null if none)"
          }
        }
      }
    },
...
  }
}